refactor modules hooks
