+++ lib/app/lift.js
@@ -84,7 +84,7 @@ function _printSuccessMsg(sails) {
       sails.log('Host          : ' + sails.getHost()); // 12 - 4 = 8 spaces
     }
     sails.log('Port           : ' + sails.config.port); // 12 - 4 = 8 spaces
-    if (!sails.modules.isEmpty()) _printLoaderMsg("Modules loaded :", sails.modules.all());
+    if (!sails.modules.isEmpty()) _printLoaderMsg("Modules loaded :", sails.modules.availables());
     sails.log('--------------------------------------------------------'.grey);
 
   }
