+++ lib/hooks/modules/index.js
@@ -1,8 +1,10 @@
 /**
  * Dependencies
  */
-var pathHelper = require('./pathHelper');
-var Err        = require('../../../errors');
+var path = require("path");
+var _    = require("lodash");
+var Err  = require('../../../errors');
+
 
 
 /**
@@ -12,43 +14,55 @@ module.exports = function(sails){
   return {
     initialize: function(cb){
       sails.modules = {
-        availables   : _availables,
-        isAvailable  : _isAvailable,
-        size         : _size,
-        isEmpty      : _isEmpty,
-        errorHandler : _errorHandler,
-        shield       : _shield
+        availables: function(){
+          return _list(sails.config.appPath, sails.config.paths.modules);
+        },
+        isAvailable: function(moduleName){
+          return _isAvailable(sails.config.appPath, sails.config.paths.modules, moduleName);
+        },
+        size: function(){
+          return _size(sails.config.appPath, sails.config.paths.modules);
+        },
+        isEmpty: function(){
+          return (_size(sails.config.appPath, sails.config.paths.modules) > 0) ? false : true;
+        },
+        errorHandler: function(moduleName){
+          Err.fatal.__ModuleMissing__(moduleName);
+          return;
+        },
+        shield: function(){
+          _.forEach(arguments, function(moduleName){
+            if(!_isAvailable(sails.config.appPath, sails.config.paths.modules, moduleName)) {
+              Err.fatal.__ModuleMissing__(moduleName);
+              return;
+            }
+          });
+        }
       };
       cb();
     }
   };
 };
 
-/**
- * Private
- */
- function _availables(){
-  return pathHelper.list(sails.config.appPath, sails.config.paths.modules);
-}
+ /**
+  * Private
+  */
+  function _list(appPath, arrayPath){
+    var list = [];
+    for (var i = arrayPath.length - 1; i >= 0; i--)
+      list.push((path.basename(arrayPath[i])));
+    // TODO: Deleted when moduleloader are fixed
+    // ensure that appPath is not in the array
+    return _.pull(list, path.basename(appPath));
+  }
 
-function _isAvailable(moduleName){
-  return pathHelper.isAvailable(sails.config.appPath, sails.config.paths.modules, moduleName);
-}
+  function _isAvailable(appPath, arrayPath, item){
+    list = _list(appPath, arrayPath);
+    result = _.indexOf(list, item);
+    if (result === -1) return false;
+    return true;
+  }
 
-function _size(){
-  return _availables().length;
-}
-
-function _isEmpty(){
-  return (_size() > 0) ? false : true;
-}
-
-function _errorHandler(moduleName){
-  Err.fatal.__ModuleMissing__(moduleName);
-  return;
-}
-
-function _shield(){
-  for (var i = arguments.length - 1; i >= 0; i--)
-    if(!_isAvailable(arguments[i])) _errorHandler(arguments[i]);
-}
+  function _size(appPath, arrayPath){
+    return _list(appPath, arrayPath).length;
+  }
